fix: implement ThemeSwitcher component for dynamic theme selection
